Move auth utils
