+++ src/main/webapp/ui/src/modules/common/utils/auth.ts
@@ -1,0 +1,15 @@
+export type TokenParams = {
+  token?: string;
+  getToken?: () => Promise<string>;
+};
+
+export const resolveToken = async ({ token, getToken }: TokenParams) => {
+  if (token) {
+	return token;
+  }
+  if (getToken) {
+	return getToken();
+  }
+  throw new Error("Token is required to perform this operation");
+};
+
