+++ src/main/webapp/ui/src/modules/stoichiometry/__tests__/utils.test.ts
@@ -1,34 +1,5 @@
-import { describe, expect, it, vi } from "vitest";
-import {
-  resolveToken,
-  toStoichiometryError,
-} from "@/modules/stoichiometry/utils";
-
-describe("resolveToken", () => {
-  it("returns direct token when provided and does not call getToken", async () => {
-    const getToken = vi.fn(async () => "callback-token");
-
-    const result = await resolveToken({ token: "direct-token", getToken });
-
-    expect(result).toBe("direct-token");
-    expect(getToken).not.toHaveBeenCalled();
-  });
-
-  it("uses getToken when direct token is not provided", async () => {
-    const getToken = vi.fn(async () => "callback-token");
-
-    const result = await resolveToken({ getToken });
-
-    expect(result).toBe("callback-token");
-    expect(getToken).toHaveBeenCalledTimes(1);
-  });
-
-  it("throws when neither token nor getToken is provided", async () => {
-    await expect(resolveToken({})).rejects.toThrow(
-      "Token is required to perform this operation",
-    );
-  });
-});
+import { describe, expect, it } from "vitest";
+import { toStoichiometryError } from "@/modules/stoichiometry/utils";
 
 describe("toStoichiometryError", () => {
   it("returns REST API error message when payload matches RestApiError schema", () => {
