+++ src/main/webapp/ui/src/modules/stoichiometry/queries.ts
@@ -5,11 +5,13 @@ import {
   StoichiometryResponseSchema,
 } from "@/modules/stoichiometry/schema";
 import {
-  resolveToken,
   STOICHIOMETRY_API_BASE_URL,
-  TokenParams,
   toStoichiometryError,
 } from "@/modules/stoichiometry/utils";
+import {
+  resolveToken,
+  TokenParams,
+} from "@/modules/common/utils/auth";
 
 export type GetStoichiometryParams = {
   stoichiometryId: number;
