+++ src/main/webapp/ui/src/modules/stoichiometry/utils.ts
@@ -7,20 +7,6 @@ import {
 
 export const STOICHIOMETRY_API_BASE_URL = "/api/v1";
 
-export type TokenParams = {
-  token?: string;
-  getToken?: () => Promise<string>;
-};
-
-export const resolveToken = async ({ token, getToken }: TokenParams) => {
-  if (token) {
-    return token;
-  }
-  if (getToken) {
-    return getToken();
-  }
-  throw new Error("Token is required to perform this operation");
-};
 
 export const toStoichiometryError = (
   data: unknown,
